Finished queries regarding creation of meal with adding of products

+++ models/productforuser.js
@@ -1,7 +1,7 @@
 "use strict";
 module.exports = (sequelize, DataTypes) => {
   const ProductForUser = sequelize.define("ProductForUser", {
-    productQuantity: DataTypes.FLOAT
+    productQuantity: DataTypes.FLOAT(15, 3)
   });
 
   return ProductForUser;
